+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/topNavFeeds/allFollows/AllFollowsFeedFlow.kt
@@ -20,56 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.topNavFeeds.allFollows
 
-import com.vitorpamplona.amethyst.model.serverList.MergedFollowListsState
-import com.vitorpamplona.amethyst.model.topNavFeeds.IFeedFlowsType
-import com.vitorpamplona.amethyst.model.topNavFeeds.IFeedTopNavFilter
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import kotlinx.coroutines.flow.FlowCollector
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.combine
-
-class AllFollowsFeedFlow(
-    val allFollows: StateFlow<com.vitorpamplona.amethyst.commons.model.serverList.MergedFollowListsState.AllFollows?>,
-    val followsRelays: StateFlow<Set<NormalizedRelayUrl>>,
-    val blockedRelays: StateFlow<Set<NormalizedRelayUrl>>,
-    val proxyRelays: StateFlow<Set<NormalizedRelayUrl>>,
-) : IFeedFlowsType {
-    fun convert(
-        allFollows: com.vitorpamplona.amethyst.commons.model.serverList.MergedFollowListsState.AllFollows?,
-        proxyRelays: Set<NormalizedRelayUrl>,
-    ): IFeedTopNavFilter =
-        if (allFollows != null) {
-            if (proxyRelays.isEmpty()) {
-                AllFollowsByOutboxTopNavFilter(
-                    authors = allFollows.authors,
-                    hashtags = allFollows.hashtags,
-                    geotags = allFollows.geotags,
-                    communities = allFollows.communities,
-                    defaultRelays = followsRelays,
-                    blockedRelays = blockedRelays,
-                )
-            } else {
-                AllFollowsByProxyTopNavFilter(
-                    authors = allFollows.authors,
-                    hashtags = allFollows.hashtags,
-                    geotags = allFollows.geotags,
-                    communities = allFollows.communities,
-                    proxyRelays = proxyRelays,
-                )
-            }
-        } else {
-            AllFollowsByOutboxTopNavFilter(
-                authors = emptySet(),
-                defaultRelays = followsRelays,
-                blockedRelays = blockedRelays,
-            )
-        }
-
-    override fun flow() = combine(allFollows, proxyRelays, ::convert)
-
-    override fun startValue(): IFeedTopNavFilter = convert(allFollows.value, proxyRelays.value)
-
-    override suspend fun startValue(collector: FlowCollector<IFeedTopNavFilter>) {
-        collector.emit(startValue())
-    }
-}
+typealias AllFollowsFeedFlow = com.vitorpamplona.amethyst.commons.model.topNavFeeds.allFollows.AllFollowsFeedFlow
