+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/topNavFeeds/allUserFollows/Kind3UserFollowsFeedFlow.kt
@@ -20,50 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.topNavFeeds.allUserFollows
 
-import com.vitorpamplona.amethyst.commons.model.nip02FollowLists.Kind3FollowListState
-import com.vitorpamplona.amethyst.model.topNavFeeds.IFeedFlowsType
-import com.vitorpamplona.amethyst.model.topNavFeeds.IFeedTopNavFilter
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import kotlinx.coroutines.flow.FlowCollector
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.combine
-
-class Kind3UserFollowsFeedFlow(
-    val allFollows: StateFlow<Kind3FollowListState.Kind3Follows>,
-    val followsRelays: StateFlow<Set<NormalizedRelayUrl>>,
-    val blockedRelays: StateFlow<Set<NormalizedRelayUrl>>,
-    val proxyRelays: StateFlow<Set<NormalizedRelayUrl>>,
-) : IFeedFlowsType {
-    fun convert(
-        allFollows: Kind3FollowListState.Kind3Follows?,
-        proxyRelays: Set<NormalizedRelayUrl>,
-    ): IFeedTopNavFilter =
-        if (allFollows != null) {
-            if (proxyRelays.isEmpty()) {
-                AllUserFollowsByOutboxTopNavFilter(
-                    authors = allFollows.authors,
-                    defaultRelays = followsRelays,
-                    blockedRelays = blockedRelays,
-                )
-            } else {
-                AllUserFollowsByProxyTopNavFilter(
-                    authors = allFollows.authors,
-                    proxyRelays = proxyRelays,
-                )
-            }
-        } else {
-            AllUserFollowsByOutboxTopNavFilter(
-                authors = emptySet(),
-                defaultRelays = followsRelays,
-                blockedRelays = blockedRelays,
-            )
-        }
-
-    override fun flow() = combine(allFollows, proxyRelays, ::convert)
-
-    override fun startValue(): IFeedTopNavFilter = convert(allFollows.value, proxyRelays.value)
-
-    override suspend fun startValue(collector: FlowCollector<IFeedTopNavFilter>) {
-        collector.emit(startValue())
-    }
-}
+typealias Kind3UserFollowsFeedFlow = com.vitorpamplona.amethyst.commons.model.topNavFeeds.allUserFollows.Kind3UserFollowsFeedFlow
